Chunk load store

+++ src/settings.rs
@@ -1,5 +1,5 @@
 // In-game settings that the use controls
 // Not for internal settings
 pub struct Settings {
-    pub render_distance: u32,
+    pub render_distance: i64,
 }
